requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
pandas-ta
openpyxl
xlrd
google-genai>=1.0.0
line-bot-sdk>=3.10.0
python-dotenv>=1.0.0
jpholiday
